pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "figma-structured-mcp"
version = "1.0.0"
description = "A powerful MCP server for interacting with the Figma API, with support for image exporting, compression, and uploading."
authors = [
    {name = "PursueLLL", email = "yululiu2018@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.8.0",
    "httpx>=0.28.0",
    "pillow>=11.0.0",
    "imagequant>=1.1.4",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
figma-structured-mcp = "figma_structured_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/figma_structured_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

